moved all states to CcdStates class
