+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/maintenance/SubmitDaCaseTest.java
@@ -12,7 +12,7 @@ import uk.gov.hmcts.reform.divorce.support.cos.CosApiClient;
 import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.DA_REQUESTED;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.DA_REQUESTED;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.DECREE_NISI_GRANTED_DATE_CCD_FIELD;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.MAKE_CASE_ELIGIBLE_FOR_DA_PETITIONER_EVENT_ID;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.STATE_CCD_FIELD;
