+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/domain/model/OrchestrationConstants.java
@@ -179,29 +179,6 @@ public class OrchestrationConstants {
     public static final String BO_WELSH_DN_RECEIVED_REVIEW = "boWelshDnReceivedReview";
     public static final String BO_WELSH_GRANT_DN_MAKE_DECISION = "boWelshGrantDnMakeDecision";
 
-    // CCD Case States
-    public static final String BO_WELSH_RESPONSE_AWAITING_REVIEW = "WelshResponseAwaitingReview";
-    public static final String AOS_AWAITING = "AosAwaiting";
-    public static final String AOS_AWAITING_SOLICITOR = "AosAwaitingSol";
-    public static final String AOS_COMPLETED = "AosCompleted";
-    public static final String AOS_OVERDUE = "AosOverdue";
-    public static final String AOS_STARTED = "AosStarted";
-    public static final String AOS_SUBMITTED_AWAITING_ANSWER = "AosSubmittedAwaitingAnswer";
-    public static final String AWAITING_DECREE_NISI = "AwaitingDecreeNisi";
-    public static final String AWAITING_LEGAL_ADVISOR_REFERRAL = "AwaitingLegalAdvisorReferral";
-    public static final String AWAITING_PRONOUNCEMENT = "AwaitingPronouncement";
-    public static final String AWAITING_CLARIFICATION = "AwaitingClarification";
-    public static final String AWAITING_REISSUE = "AwaitingReissue";
-    public static final String DEFENDED = "DefendedDivorce";
-    public static final String DN_PRONOUNCED = "DNPronounced";
-    public static final String AWAITING_DA = "AwaitingDecreeAbsolute";
-    public static final String DN_REFUSED = "DNisRefused";
-    public static final String WELSH_LA_DECISION = "WelshLADecision";
-    public static final String AWAITING_ADMIN_CLARIFICATION = "AwaitingAdminClarification";
-    public static final String DA_REQUESTED = "DARequested";
-    public static final String DA_OVERDUE = "DAOverdue";
-    public static final String DIVORCE_GRANTED = "DivorceGranted";
-
     // CCD Co-Respondent Fields
     public static final String CO_RESP_LINKED_TO_CASE = "CoRespLinkedToCase";
     public static final String CO_RESP_LINKED_TO_CASE_DATE = "CoRespLinkedToCaseDate";
@@ -340,8 +317,6 @@ public class OrchestrationConstants {
 
     //Bulk print letter types
     public static final String DA_GRANTED_OFFLINE_PACK_RESPONDENT = "da-granted-offline-pack-respondent";
-    public static final String COE_OFFLINE_PACK_RESPONDENT = "coe-offline-pack-respondent";
-    public static final String COST_ORDER_OFFLINE_PACK_CO_RESPONDENT = "co-respondent-offline-pack-respondent";
 
     public static final String DOCUMENT_TYPE_COE = "coe";
     public static final String DOCUMENT_TYPE_OTHER = "other";
@@ -438,5 +413,4 @@ public class OrchestrationConstants {
 
     // Bulk Scan
     public static final String CASE_TYPE_ID = "DIVORCE";
-
 }
